requirements.txt
python-dotenv
langchain-openai
langchain-groq
langchain
pandas
tqdm